Feat: organizar proyecto
