+++ src/entities/usuario.py
@@ -2,22 +2,25 @@ import uuid
 
 from sqlalchemy import Column, DateTime, ForeignKey, String
 from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import relationship
 
 from database import Base
 
 
-class Paciente(Base):
+class Usuario(Base):
     """
-    Modelo de paciente
+    Modelo de usuario
     """
 
-    __tablename__ = "pacientes"
-    idPaciente = Column(
+    __tablename__ = "usuarios"
+    id_usuario = Column(
         UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, index=True
     )
-    nombrePaciente = Column(String, index=True)
-    correoPaciente = Column(String, index=True)
+    primer_nombre = Column(String, index=True, nullable=False)
+    segundo_nombre = Column(String, index=True, nullable=True)
+    primer_apellido = Column(String, index=True, nullable=False)
+    segundo_apellido = Column(String, index=True, nullable=True)
+    correo_usuario = Column(String, unique=True, index=True, nullable=False)
+    contrasena_usuario = Column(String, nullable=False)
 
     # Campos de auditoría
     id_usuario_creacion = Column(
@@ -28,8 +31,3 @@ class Paciente(Base):
     )
     fecha_creacion = Column(DateTime, index=True)
     fecha_actualizacion = Column(DateTime, index=True)
-
-    # Relaciones
-    citas = relationship("Cita", back_populates="paciente")
-    diagnosticos = relationship("Diagnostico", back_populates="paciente")
-    facturas = relationship("Factura", back_populates="paciente")
